pyproject.toml
[project]
name = "MegaMegaschool"
version = "0.1.0"
description = ""
authors = [
    {name = "Ivan",email = "ivan.golovachev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=1.60.2,<2.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "tenacity (>=9.0.0,<10.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "googlesearch-python (>=1.3.0,<2.0.0)",
    "faker (>=35.2.0,<36.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
